+++ src/test/java/com/hudhaifa/sortframework/sort/BubbleSortTest.java
@@ -13,32 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hudhaifa.sortframework.algo.sort;
+package com.hudhaifa.sortframework.sort;
 
-import com.hudhaifa.sortframework.algo.sort.MergeSort;
-import com.hudhaifa.sortframework.util.ArrayUtil;
-import java.util.Arrays;
+import com.hudhaifa.sortframework.sort.BubbleSort;
 import org.junit.After;
-import static org.junit.Assert.assertArrayEquals;
 import org.junit.Before;
 import org.junit.Test;
+
+import static org.junit.Assert.assertArrayEquals;
 
 /**
  *
  * @author Hudhaifa Shatnawi <hudhaifa.shatnawi@gmail.com>
- * @version 1.0, Oct 13, 2013 - 10:25:32 PM
+ * @version 1.0, Oct 11, 2013 - 8:45:27 AM
  */
-public class MergeSortTest {
+public class BubbleSortTest {
 
-    public MergeSortTest() {
+    public BubbleSortTest() {
     }
 
     @Before
     public void setUp() {
-        instance = new MergeSort();
-        a = ArrayUtil.random(100_000_000);
-        b = ArrayUtil.cloneArray(a);
-        Arrays.sort(b);
+        instance = new BubbleSort();
     }
 
     @After
@@ -47,7 +43,7 @@ public class MergeSortTest {
     }
 
     /**
-     * Test of sort method, of class InsertionSort.
+     * Test of sort method, of class BubbleSort.
      */
     @Test
     public void testSort() {
@@ -59,12 +55,10 @@ public class MergeSortTest {
         System.out.println("Array size: " + a.length);
         System.out.println("# Comparisons: " + instance.getComparisons());
         System.out.println("# Swaps: " + instance.getSwaps());
-        System.out.println("# Compexity: " + instance.getTimeCompleixty());
 
         assertArrayEquals(a, b);
     }
-
-    private MergeSort instance;
-    private int[] a;
-    private int[] b;
+    private BubbleSort instance;
+    private int[] a = {2, 1, 5, 0, 5, 4};
+    private int[] b = {0, 1, 2, 4, 5, 5};
 }
