+++ src/test/java/com/hudhaifa/sortframework/sort/ShuffleSortTest.java
@@ -13,11 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hudhaifa.sortframework.algo.sort;
+package com.hudhaifa.sortframework.sort;
 
-import com.hudhaifa.sortframework.algo.sort.InsertionSort;
-import com.hudhaifa.sortframework.util.ArrayUtil;
-import java.util.Arrays;
+import com.hudhaifa.sortframework.sort.ShuffleSort;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -27,19 +25,16 @@ import static org.junit.Assert.assertArrayEquals;
 /**
  *
  * @author Hudhaifa Shatnawi <hudhaifa.shatnawi@gmail.com>
- * @version 1.0, Oct 11, 2013 - 8:19:58 AM
+ * @version 1.0, Oct 11, 2013 - 8:45:27 AM
  */
-public class InsertionSortTest {
+public class ShuffleSortTest {
 
-    public InsertionSortTest() {
+    public ShuffleSortTest() {
     }
 
     @Before
     public void setUp() {
-        instance = new InsertionSort();
-        a = ArrayUtil.random(100);
-        b = ArrayUtil.cloneArray(a);
-        Arrays.sort(b);
+        instance = new ShuffleSort();
     }
 
     @After
@@ -48,7 +43,7 @@ public class InsertionSortTest {
     }
 
     /**
-     * Test of sort method, of class InsertionSort.
+     * Test of sort method, of class BubbleSort.
      */
     @Test
     public void testSort() {
@@ -60,28 +55,10 @@ public class InsertionSortTest {
         System.out.println("Array size: " + a.length);
         System.out.println("# Comparisons: " + instance.getComparisons());
         System.out.println("# Swaps: " + instance.getSwaps());
-        System.out.println("# Compexity: " + instance.getTimeCompleixty());
 
         assertArrayEquals(a, b);
     }
-
-    /**
-     * Test of sort method, of class InsertionSort.
-     */
-    @Test
-    public void testSort2() {
-        System.out.println("Recursive sort");
-
-        instance.reset(a);
-        instance.sort(a.length);
-
-        System.out.println("Array size: " + a.length);
-        System.out.println("# Comparisons: " + instance.getComparisons());
-        System.out.println("# Swaps: " + instance.getSwaps());
-
-        assertArrayEquals(a, b);
-    }
-    private InsertionSort instance;
-    private int[] a;
-    private int[] b;
+    private ShuffleSort instance;
+    private int[] a = {2, 1, 5, 0, 5, 4};
+    private int[] b = {0, 1, 2, 4, 5, 5};
 }
